+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- method: `ModelHandlerTrait::model_execute_splat()`
 
 ## [1.5.1] - 2018-12-11
 ### Added
@@ -12,7 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix missing assignation of `$this->currentEndpoint` in `OdooClient::getService()` when service already set
-
 
 ## [1.5.0] - 2018-12-10